Move useConfirmModal on store/
