+++ conversation/frontend/src/providers/index.tsx
@@ -7,7 +7,7 @@ import {
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { ERROR_CODE } from 'edifice-ts-client';
 import { ReactNode } from 'react';
-import { useConfirmModalStore } from '~/hooks/useConfirmModalStore';
+import { useConfirmModalStore } from '~/store';
 
 export const queryClient = new QueryClient({
   queryCache: new QueryCache({
