+++ conversation/frontend/src/store/index.ts
@@ -6,3 +6,4 @@
  */
 export * from './rights';
 export * from './actions';
+export * from './confirmModal';
